+++ adapters/__init__.py
@@ -23,6 +23,7 @@ from adapters.lumi.WXKG02LM import WXKG02LM
 from adapters.lumi.WXKG03LM import WXKG03LM
 from adapters.lumi.WXKG11LM import WXKG11LM
 from adapters.lumi.WXKG12LM import WXKG12LM
+from adapters.lumi.ZNCZ02LM import ZNCZ02KM
 from adapters.samsung.sensor_arrival import SensorArrival
 from adapters.philips.hue_dimmer_switch import HueDimmerSwitch
 from adapters.philips.hue_motion_sensor import HueMotionSensor
@@ -225,7 +226,7 @@ adapter_by_model = {
     'WXKG03LM': WXKG03LM,               # Xiaomi Aqara single key wireless wall switch
     'WXKG11LM': WXKG11LM,               # Xiaomi Aqara wireless switch
     'WXKG12LM': WXKG12LM,               # Xiaomi Aqara wireless switch (with gyroscope)
-    'ZNCZ02LM': Plug,                   # Xiaomi Mi power plug ZigBee
+    'ZNCZ02LM': ZNCZ02KM,               # Xiaomi Mi power plug ZigBee
     'ZNLDP12LM': DimmableCtBulbAdapter, # Xiaomi Aqara smart LED bulb
     # iCasa
     'ICZB-IW11D': DimmableBulbAdapter,  # iCasa Zigbee 3.0 Dimmer